adding a comparison picture of final traj to gif_generator

+++ plots/gifs.py
@@ -131,8 +131,8 @@ def trajectory_gif(model, inputs, targets, timesteps, dpi=200, alpha=0.9,
         plt.savefig(base_filename + "{}.png".format(t),
                     format='png', dpi=dpi, bbox_inches='tight')
         # Save only 3 frames (.pdf for paper)
-        if t in [0, interp_time//5, interp_time//2, interp_time-1]:
-            plt.savefig(base_filename + "{}.pdf".format(t), format='pdf', bbox_inches='tight')
+        # if t in [0, interp_time//5, interp_time//2, interp_time-1]:
+        #     plt.savefig(base_filename + "{}.pdf".format(t), format='pdf', bbox_inches='tight')
         plt.clf()
         plt.close()
 
@@ -140,7 +140,7 @@ def trajectory_gif(model, inputs, targets, timesteps, dpi=200, alpha=0.9,
     for i in range(interp_time):
         img_file = base_filename + "{}.png".format(i)
         imgs.append(imageio.imread(img_file))
-        os.remove(img_file) 
+        if i not in [0, interp_time//5, interp_time//2, interp_time-1]: os.remove(img_file) 
     imageio.mimwrite(filename, imgs)
 
 def mnist_gif(model, inputs, timesteps, component, filename='mnist.gif'):
